requirements.txt
tensorflow==2.14.0
keras==3.14.0
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow==10.1.0
Pillow==10.2.0
scipy==1.12.0